@@ -1,7 +1,7 @@
-import { Injectable } from '@angular/core';
-import { ApiService } from './api.service'
+import {Injectable} from '@angular/core';
+import {ApiService} from './api.service';
 
-import { environment } from '../../_environments/environment';
+import {environment} from '../../_environments/environment';
 
 @Injectable({
   providedIn: 'root'
@@ -13,76 +13,70 @@ export class QuestionService {
 	}
 
   getLineItemLevelAssociations(questionId) {
-    let url = environment.apiUrl + "/api/question/" + questionId + "/lineitem/levels";
+    const url = environment.apiUrl + '/api/question/' + questionId + '/lineitem/levels';
 
-    let rtn = new Promise(
+    return new Promise(
       (resolve, reject) => {
-      this._apiService.getUnsecuredAPI(url).subscribe(
-        (data) => {
-          resolve(data);
-        }, (err) => {
-          reject(err);
-        });
-      }
-    );
-
-    return rtn;
-  }
-
-  setLineItemLevelAssociations(questionId, lilvassociations) {
-    let url = environment.apiUrl + "/api/question/" + questionId + "/lineitem/levels";
-
-    let data = '';
-
-    for (var x=0; x < lilvassociations.length; x++) {
-      if (x > 0) data += '&';
-
-      data += 'liId' + x + '=' + lilvassociations[x][0] + '&liVal' + x + '=' + lilvassociations[x][1];
-    }
-
-    if (data.length > 0)
-      data += '&count=' + lilvassociations.length;
-
-    let rtn = new Promise(
-      (resolve, reject) => {
-      this._apiService.postUnsecuredAPI(url, data).subscribe(
-        (data) => {
-          resolve(data);
-        }, (err) => {
-          reject(err);
-        });
-      }
-    );
-
-    return rtn;
-  }
-
-  getAll() {
-      let url = environment.apiUrl + "/api/question/all";
-
-      let rtn = new Promise(
-        (resolve, reject) => {
         this._apiService.getUnsecuredAPI(url).subscribe(
           (data) => {
             resolve(data);
           }, (err) => {
             reject(err);
           });
+      }
+    );
+  }
+
+  setLineItemLevelAssociations(questionId, lilvassociations) {
+    const url = environment.apiUrl + '/api/question/' + questionId + '/lineitem/levels';
+
+    let data = '';
+
+    for (let x=0; x < lilvassociations.length; x++) {
+      if (x > 0) {data += '&';}
+
+      data += 'liId' + x + '=' + lilvassociations[x][0] + '&liVal' + x + '=' + lilvassociations[x][1];
+    }
+
+    if (data.length > 0)
+      {data += '&count=' + lilvassociations.length;}
+
+    return new Promise(
+      (resolve, reject) => {
+        this._apiService.postUnsecuredAPI(url, data).subscribe(
+          (_data) => {
+            resolve(_data);
+          }, (err) => {
+            reject(err);
+          });
+      }
+    );
+  }
+
+  getAll() {
+      const url = environment.apiUrl + '/api/question/all';
+
+    return new Promise(
+        (resolve, reject) => {
+          this._apiService.getUnsecuredAPI(url).subscribe(
+            (data) => {
+              resolve(data);
+            }, (err) => {
+              reject(err);
+            });
         }
       );
-
-      return rtn;
   }
 
   getQuestionById(id) {
-      let url = environment.apiUrl + "/api/question/" + id;
+      const url = environment.apiUrl + '/api/question/' + id;
 
-      let rtn = new Promise(
+      const rtn = new Promise(
         (resolve, reject) => {
         this._apiService.getUnsecuredAPI(url).subscribe(
           (data) => {
-            console.log("Call to getQuestionById(" + id + ") returned")
-            console.log(data)
+            console.log('Call to getQuestionById(' + id + ') returned');
+            console.log(data);
             resolve(data);
           }, (err) => {
             reject(err);
@@ -94,26 +88,24 @@ export class QuestionService {
   }
 
   getByLineItemAndLevel(lineItemId, levelIdx) {
-      let url = environment.apiUrl + "/api/question/" + lineItemId + "/" + levelIdx;
+      const url = environment.apiUrl + '/api/question/' + lineItemId + '/' + levelIdx;
 
-      let rtn = new Promise(
+    return new Promise(
         (resolve, reject) => {
-        this._apiService.getUnsecuredAPI(url).subscribe(
-          (data) => {
-            resolve(data);
-          }, (err) => {
-            reject(err);
-          });
+          this._apiService.getUnsecuredAPI(url).subscribe(
+            (data) => {
+              resolve(data);
+            }, (err) => {
+              reject(err);
+            });
         }
       );
-
-      return rtn;
   }
 
 	getUserHistoryForQuestion(userId, questionId) {
-      let url = environment.apiUrl + "/api/user/" + userId + "/question/" + questionId + "/history";
+      const url = environment.apiUrl + '/api/user/' + userId + '/question/' + questionId + '/history';
 
-      let rtn = new Promise(
+      const rtn = new Promise(
         (resolve, reject) => {
         this._apiService.getUnsecuredAPI(url).subscribe(
           (data) => {
@@ -128,17 +120,17 @@ export class QuestionService {
 	}
 
   setSessionScore(userId, questionId, sessionId, dataObj) {
-      let url = environment.apiUrl + "/api/user/" + userId + "/question/" + questionId + "/history";
-      let data = "sessionId=" + sessionId + "&score=" + dataObj["score"];
+      const url = environment.apiUrl + '/api/user/' + userId + '/question/' + questionId + '/history';
+      let data = 'sessionId=' + sessionId + '&score=' + dataObj.score;
 
-      if (dataObj["comment"]) 
-        data += "&comment=" + dataObj["comment"];
+      if (dataObj.comment)
+        {data += '&comment=' + dataObj.comment;}
 
-      let rtn = new Promise(
+      const rtn = new Promise(
         (resolve, reject) => {
         this._apiService.postUnsecuredAPI(url, data).subscribe(
-          (data) => { 
-            console.log("Question Session Score Updated!");
+          (data) => {
+            console.log('Question Session Score Updated!');
             console.log(data);
 
             resolve(data);
@@ -151,15 +143,15 @@ export class QuestionService {
   }
 
   save(question, lilvassociations) {
-    let url = environment.apiUrl + "/api/question/save";
+    const url = environment.apiUrl + '/api/question/save';
 
     return new Promise(
       (resolve, reject) => {
-        this._apiService.postUnsecuredAPI2(url, {question: question, lilvassociations: lilvassociations}).subscribe(
+        this._apiService.postUnsecuredAPI2(url, {question, lilvassociations}).subscribe(
           (data) => {
-            resolve(data)
+            resolve(data);
           }, (err) => {
-            reject(err)
+            reject(err);
           });
       });
   }
