@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import { ApiService } from './api.service'
-import { UserService } from './user.service'
+import { ApiService } from './api.service';
+import { UserService } from './user.service';
 
 import { environment } from '../../_environments/environment';
 
@@ -15,13 +15,13 @@ export class TechProfileAPIService {
   }
 
   get(techProfileId) {
-    let url = environment.apiUrl + "/api/techprofile/" + techProfileId;
+    const url = environment.apiUrl + '/api/techprofile/' + techProfileId;
 
-  	let rtn = new Promise(
+  	const rtn = new Promise(
   		(resolve, reject) => {
   			this._apiService.get(url).subscribe(
-  				(data) => { 
-  					console.log("getTechProfile API call returned");
+  				(data) => {
+  					console.log('getTechProfile API call returned');
   					console.log(data);
 
   					resolve(data);
@@ -34,14 +34,14 @@ export class TechProfileAPIService {
   }
 
   getScores(userId) {
-    console.log("calling to get TechProfile scores for [" + userId + "] ");
-  	let url = environment.apiUrl + "/api/user/" + userId + "/techprofile/scores";
+    console.log('calling to get TechProfile scores for [' + userId + '] ');
+  	const url = environment.apiUrl + '/api/user/' + userId + '/techprofile/scores';
 
-  	let rtn = new Promise(
+  	const rtn = new Promise(
   		(resolve, reject) => {
   			this._apiService.get(url).subscribe(
-  				(data) => { 
-  					console.log("get TechProfile scores for [" + userId + "] API call returned");
+  				(data) => {
+  					console.log('get TechProfile scores for [' + userId + '] API call returned');
   					console.log(data);
 
   					resolve(data);
@@ -50,45 +50,45 @@ export class TechProfileAPIService {
   				});
   		});
 
-  	return rtn;  
+  	return rtn;
   }
 
   saveScores(userId, scores) {
-  	let url = environment.apiUrl + "/api/user/" + userId + "/techprofile/scores";
+  	const url = environment.apiUrl + '/api/user/' + userId + '/techprofile/scores';
 
-  	let data = this.JSON_to_URLEncoded(scores);
+  	const data = this.JSON_to_URLEncoded(scores);
 
-  	let rtn = new Promise(
+  	const rtn = new Promise(
   		(resolve, reject) => {
   			this._apiService.post(url, data).subscribe(
-  				(data) => { 
-  					console.log("POST TechProfile scores for [" + userId + "] API call returned");
-  					console.log(data);
+  				(_data) => {
+  					console.log('POST TechProfile scores for [' + userId + '] API call returned');
+  					console.log(_data);
 
-  					resolve(data);
+  					resolve(_data);
   				}, (err) => {
   					reject(err);
   				});
   		});
 
-  	return rtn;  
+  	return rtn;
   }
 
   addTopic(name) {
-  	let url = environment.apiUrl + "/api/techprofile/topics/new"
+  	const url = environment.apiUrl + '/api/techprofile/topics/new';
 
-  	let data = "topicName="+name
+  	const data = 'topicName='+name;
 
-  	let rtn = new Promise(
+  	const rtn = new Promise(
   		(resolve, reject) => {
   			this._apiService.post(url, data).subscribe(
-  				(data) => {
-  					console.log("POST addTopic [" + name + "] API call returned")
-  					console.log(data)
+  				(_data) => {
+  					console.log('POST addTopic [' + name + '] API call returned');
+  					console.log(_data);
 
-  					resolve(data)
+  					resolve(_data);
   				}, (err) => {
-  					reject(err)
+  					reject(err);
   				});
   		});
 
@@ -96,28 +96,28 @@ export class TechProfileAPIService {
   }
 
   addLineItem(topicId, name) {
-  	return this.addLineItemWithDescriptions(topicId, name, "level 0 desc", "level 1 desc", "level 2 desc", "level 3 desc");
+  	return this.addLineItemWithDescriptions(topicId, name, 'level 0 desc', 'level 1 desc', 'level 2 desc', 'level 3 desc');
   }
 
   addLineItemWithDescriptions(topicId, name, l0description, l1description, l2description, l3description) {
-  	let url = environment.apiUrl + "/api/techprofile/topics/" + topicId + "/lineitem/new"
+  	const url = environment.apiUrl + '/api/techprofile/topics/' + topicId + '/lineitem/new';
 
-  	let data = "lineItemName="+name
-  		+"&l0description="+l0description
-  		+"&l1description="+l1description
-  		+"&l2description="+l2description
-  		+"&l3description="+l3description;
+  	const data = 'lineItemName='+name
+  		+'&l0description='+l0description
+  		+'&l1description='+l1description
+  		+'&l2description='+l2description
+  		+'&l3description='+l3description;
 
-  	let rtn = new Promise(
+  	const rtn = new Promise(
   		(resolve, reject) => {
   			this._apiService.post(url, data).subscribe(
-  				(data) => {
-  					console.log("POST addLineItem [" + name + "] API call returned")
-  					console.log(data)
+  				(_data) => {
+  					console.log('POST addLineItem [' + name + '] API call returned');
+  					console.log(_data);
 
-  					resolve(data)
+  					resolve(_data);
   				}, (err) => {
-  					reject(err)
+  					reject(err);
   				});
   		});
 
@@ -125,25 +125,25 @@ export class TechProfileAPIService {
   }
 
   updateLineItemWithDescriptions(lineItem) {
-  	let url = environment.apiUrl + "/api/techprofile/lineitem/" + lineItem["id"];
-    let rtn = new Promise((resolve, reject) => { reject("invalid data") });
+  	const url = environment.apiUrl + '/api/techprofile/lineitem/' + lineItem.id;
+    let rtn = new Promise((resolve, reject) => { reject('invalid data'); });
 
-  	if (lineItem["name"] !== undefined 
-        && lineItem["l0Description"] !== undefined
-        && lineItem["l1Description"] !== undefined
-        && lineItem["l2Description"] !== undefined
-        && lineItem["l3Description"] !== undefined) {
+  	if (lineItem.name !== undefined
+        && lineItem.l0Description !== undefined
+        && lineItem.l1Description !== undefined
+        && lineItem.l2Description !== undefined
+        && lineItem.l3Description !== undefined) {
 
         	rtn = new Promise(
         		(resolve, reject) => {
-        			this._apiService.postUnsecuredAPI2(url, {lineItem: lineItem}).subscribe(
-        				(data) => {
-        					console.log("POST updateLineItem [" + lineItem['id'] + "] API call returned")
-        					console.log(data)
+        			this._apiService.postUnsecuredAPI2(url, {lineItem}).subscribe(
+        				(_data) => {
+        					console.log('POST updateLineItem [' + lineItem.id + '] API call returned');
+        					console.log(_data);
 
-        					resolve(data)
+        					resolve(_data);
         				}, (err) => {
-        					reject(err)
+        					reject(err);
         				});
         		});
         }
@@ -152,20 +152,20 @@ export class TechProfileAPIService {
   }
 
   updateTopic(topic) {
-    let url = environment.apiUrl + "/api/techprofile/topic/" + topic["id"]
-    let rtn = new Promise((resolve, reject) => { reject("invalid data") })
+    const url = environment.apiUrl + '/api/techprofile/topic/' + topic.id;
+    let rtn = new Promise((resolve, reject) => { reject('invalid data'); });
 
-    if (topic['name'] !== undefined) {
+    if (topic.name !== undefined) {
       rtn = new Promise(
         (resolve, reject) => {
-              this._apiService.postUnsecuredAPI2(url, {topic: topic}).subscribe(
-                (data) => {
-                  console.log("POST updateTopic [" + topic['id'] + "] API call returned")
-                  console.log(data)
+              this._apiService.postUnsecuredAPI2(url, {topic}).subscribe(
+                (_data) => {
+                  console.log('POST updateTopic [' + topic.id + '] API call returned');
+                  console.log(_data);
 
-                  resolve(data)
+                  resolve(_data);
                 }, (err) => {
-                  reject(err)
+                  reject(err);
                 });
             });
         }
@@ -174,40 +174,41 @@ export class TechProfileAPIService {
   }
 
   saveSequenceInfo(arr) {
-    let url = environment.apiUrl + "/api/techprofile/sequences"
-    let rtn = new Promise(
+    const url = environment.apiUrl + '/api/techprofile/sequences';
+    const rtn = new Promise(
       (resolve, reject) => {
-        this._apiService.postUnsecuredAPI2(url, {arr: arr}).subscribe(
-          (data) => {
-            console.log("POST sequence info API call returned")
-            console.log(data)
-            resolve(data)
+        this._apiService.postUnsecuredAPI2(url, {arr}).subscribe(
+          (_data) => {
+            console.log('POST sequence info API call returned');
+            console.log(_data);
+            resolve(_data);
           }, (err) => {
             reject(err);
           }
-        )
+        );
       }
-    )
+    );
 
-    return rtn
+    return rtn;
   }
 
+  // eslint-disable-next-line @typescript-eslint/naming-convention
 	JSON_to_URLEncoded(scores){
-		var list = '';
-	
-		var ctr = 0;
+		let list = '';
+
+		let ctr = 0;
 		scores.map((score) => {
-			list += "userId"+ctr+"="+score.userId;
-			list += "&techProfileLineItemId"+ctr+"="+score.techProfileLineItemId;
-			list += "&techProfileLineItemScore"+ctr+"="+score.techProfileLineItemScore;
+			list += 'userId'+ctr+'='+score.userId;
+			list += '&techProfileLineItemId'+ctr+'='+score.techProfileLineItemId;
+			list += '&techProfileLineItemScore'+ctr+'='+score.techProfileLineItemScore;
 
 			if (ctr+1 < scores.length)
-				list += "&";
+				{list += '&';}
 
 			ctr++;
-		})
+		});
 
-		list += "&count="+ (ctr);
+		list += '&count='+ (ctr);
 
 		return list;
 	}
