@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-import { TechProfileAPIService } from './tech-profile-api.service'
+import { TechProfileAPIService } from './tech-profile-api.service';
 
 
 @Injectable({
@@ -10,14 +10,14 @@ export class UserTechProfileModelService {
 
 	userScores = undefined;
 	userId = undefined;
-	dirty = false;	
+	dirty = false;
 
 	constructor(protected _techProfileAPI: TechProfileAPIService) {
-		
+
 	}
 
 	init(userId) {
-		let self = this;
+		const self = this;
 
 		if (userId !== self.userId) {
 
@@ -25,27 +25,27 @@ export class UserTechProfileModelService {
 
 			self._techProfileAPI.getScores(userId).then((scores) => {
 				self.userScores = scores;
-			})
+			});
 		} else {
-			console.log("UserTechProfileModelService did NOT init. Not sure if thats a good or bad thing")
+			console.log('UserTechProfileModelService did NOT init. Not sure if thats a good or bad thing');
 		}
 	}
 
 	waitingPromise() {
-		let self = this;
-		return new Promise((resolve, reject) => {
+		const self = this;
+		return new Promise<void>((resolve, reject) => {
 
-			function to() {
+      const to = () => {
 				setTimeout(() => {
 					if (self.isUserScoresAvailable())
-						resolve();
+						{resolve();}
 					else
-						to();
+						{to();}
 				}, 600);
-			}
+			};
 
 			to();
-		})
+		});
 	}
 
 	isDirty() {
@@ -61,30 +61,30 @@ export class UserTechProfileModelService {
 	}
 
 	getScore(lineItemId) {
-		let rtn = undefined;
+		let rtn;
 
 		if (this.userScores) {
-			let score = this.userScores.find((s) => { return s["techProfileLineItemId"] === lineItemId; });
+			const score = this.userScores.find((s) => s.techProfileLineItemId === lineItemId);
 
-			rtn = !!score ? score["techProfileLineItemScore"] : undefined;
+			rtn = !!score ? score.techProfileLineItemScore : undefined;
 		}
 
 		return rtn;
 	}
 
 	clearScore(lineItemId) {
-		this.userScores = this.userScores.filter((s) => { return s["techProfileLineItemId"] !== lineItemId; });
+		this.userScores = this.userScores.filter((s) => s.techProfileLineItemId !== lineItemId);
 	}
 
 	setLineItemScore(lineItemId, idx) {
-		let self = this;
+		const self = this;
 		return new Promise((resolve, reject) => {
 			if (self.userScores) {
-				let score = self.userScores.find((s) => { return s["techProfileLineItemId"] === lineItemId; });
-				let prevScore = Object.assign({}, score);
+				let score = self.userScores.find((s) => s.techProfileLineItemId === lineItemId);
+				const prevScore = Object.assign({}, score);
 
 				if (score) {
-					score["techProfileLineItemScore"] = idx;
+					score.techProfileLineItemScore = idx;
 				} else {
 					score = {
 						userId: self.userId,
@@ -96,21 +96,21 @@ export class UserTechProfileModelService {
 				}
 
 				self.setDirty();
-				resolve({prevScore: prevScore, newScore: score});
+				resolve({prevScore, newScore: score});
 			} else {
 				resolve(undefined);
 			}
-		})
-	}	
+		});
+	}
 
 	save() {
-		let self = this;
+		const self = this;
 		return new Promise((resolve, reject) => {
 			if (self.isDirty()) {
-				self._techProfileAPI.saveScores(self.userId, self.userScores).then((data) => {
+				self._techProfileAPI.saveScores(self.userId, self.userScores).then((_data) => {
 					self.dirty = false;
-					resolve(data);
-				})
+					resolve(_data);
+				});
 			} else {
 				resolve(undefined);
 			}
