@@ -12,12 +12,12 @@ import { HomePage } from './home.page';
     FormsModule,
     IonicModule,
     RouterModule.forChild([
-      {
-        path: '',
-        component: HomePage
-      }
-    ])
-  ],
-  declarations: [HomePage]
+    {
+      path: '',
+      component: HomePage
+    }
+  ])
+],
+declarations: [HomePage]
 })
 export class HomePageModule {}
