@@ -3,7 +3,7 @@ import { Router, ActivatedRoute, ParamMap } from '@angular/router';
 import { Location } from '@angular/common';
 
 import { ModelService } from './_services/model.service';
-import { FunctionPromiseService } from 'savvato-javascript-services'
+import { FunctionPromiseService } from 'savvato-javascript-services';
 
 import { environment } from '../../_environments/environment';
 
@@ -14,7 +14,7 @@ import { environment } from '../../_environments/environment';
 })
 export class TechProfileQuestionPage implements OnInit {
 
-	funcKey = "tpqp-controller";
+	funcKey = 'tpqp-controller';
 
 	constructor(private _location: Location,
 		    	private _router: Router,
@@ -26,28 +26,23 @@ export class TechProfileQuestionPage implements OnInit {
 	}
 
 	ngOnInit() {
-		let self = this;
+		const self = this;
 
 		self._modelService._init();
 
-		self._functionPromiseService.initFunc(self.funcKey, () => {
-			return new Promise((resolve, reject) => {
+		self._functionPromiseService.initFunc(self.funcKey, () => new Promise((resolve, reject) => {
 				resolve({
-					getEnv: () => {
-						return environment;
-					},
-					getColorMeaningString: () => {
-						return "Select a skill and level for this new question. White means there are no questions associated with this skill level. Shades of gray, the closer you get to dark, indicate the more questions, relatively speaking, for that skill level."
-					},
+					getEnv: () => environment,
+					getColorMeaningString: () => 'Select a skill and level for this new question. White means there are no questions associated with this skill level. Shades of gray, the closer you get to dark, indicate the more questions, relatively speaking, for that skill level.',
 					getBackgroundColor: (lineItem, idx) => {
-							let count = this._modelService.getQuestionCountForCell(lineItem['id'], idx);
-							let max = this._modelService.getHighestQuestionCountForAnyCell();
+							const count = this._modelService.getQuestionCountForCell(lineItem.id, idx);
+							const max = this._modelService.getHighestQuestionCountForAnyCell();
 
-							let shadesOfGray = ["#FFFFFF","#E0E0E0","#D0D0D0","#C0C0C0","#B0B0B0","#A0A0A0","#909090","#808080","#707070","#606060"]
+							const shadesOfGray = ['#FFFFFF','#E0E0E0','#D0D0D0','#C0C0C0','#B0B0B0','#A0A0A0','#909090','#808080','#707070','#606060'];
 
 							if (count && max) {
-								let p = this._modelService.getPercentileForTheNumberOfQuestionsForThisCell(lineItem['id'], idx);
-								let rtn = undefined;
+								const p = this._modelService.getPercentileForTheNumberOfQuestionsForThisCell(lineItem.id, idx);
+								let rtn;
 
 								if (p) {
 									rtn = shadesOfGray[p - 1];
@@ -56,19 +51,18 @@ export class TechProfileQuestionPage implements OnInit {
 								return rtn;
 							}
 
-							return "white";
+							return 'white';
 					},
 					onLxDescriptionClick: (lineItem, idx) => {
-						this._router.navigate(['/question-list/' + lineItem['id'] + '/' + idx]);
+						this._router.navigate(['/question-list/' + lineItem.id + '/' + idx]);
 					}
 				});
-			});
-		})
+			}));
 
 	}
 
 	getDtimTechprofileComponentController() {
-		return this._functionPromiseService.waitAndGet(this.funcKey, this.funcKey, { })
+		return this._functionPromiseService.waitAndGet(this.funcKey, this.funcKey, { });
 	}
 
 	onNewQuestionBtnClicked(q) {
