@@ -12,7 +12,7 @@ import { TechProfileModelService } from '../_services/tech-profile-model.service
 export class TechProfileTopicEditPage implements OnInit {
 
 	dirty = false;
-	topic = {id: -1, name: ''}
+	topic = {id: -1, name: ''};
 
 	constructor(private _location: Location,
 			    private _router: Router,
@@ -22,27 +22,27 @@ export class TechProfileTopicEditPage implements OnInit {
 	}
 
 	ngOnInit() {
-		let self = this;
+		const self = this;
 
 		self._techProfileModelService._init();
 
 		self._route.params.subscribe((params) => {
-			let topicId = params['topicId'] * 1;
+			const topicId = params.topicId * 1;
 
 			if (topicId) {
 				self._techProfileModelService.waitingPromise().then(() => {
-					self.topic = self._techProfileModelService.getTechProfileTopicById(topicId)
-				})
+					self.topic = self._techProfileModelService.getTechProfileTopicById(topicId);
+				});
 			}
-		})
+		});
 	}
 
-	onBackBtnClicked() { 
-		let self = this;
+	onBackBtnClicked() {
+		const self = this;
 		if (self.isDirty()) {
 			self._techProfileModelService.updateTechProfileTopic(self.topic).then((data) => {
 				self._location.back();
-			})
+			});
 		} else {
 			self._location.back();
 		}
