@@ -3,7 +3,7 @@ import { Router, ActivatedRoute, ParamMap } from '@angular/router';
 import { Location } from '@angular/common';
 
 import { AlertService } from '../_services/alert.service';
-import { FunctionPromiseService } from 'savvato-javascript-services'
+import { FunctionPromiseService } from 'savvato-javascript-services';
 import { TechProfileAPIService } from '../_services/tech-profile-api.service';
 import { UserTechProfileModelService } from '../_services/user-tech-profile-model.service';
 import { UserService } from '../_services/user.service';
@@ -23,7 +23,7 @@ export class UserTechProfilePage implements OnInit {
   	user = undefined;
   	techProfile = undefined;
 
-  	funcKey = "utp-controller";
+  	funcKey = 'utp-controller';
 
 	constructor(private _location: Location,
 		    private _router: Router,
@@ -37,41 +37,35 @@ export class UserTechProfilePage implements OnInit {
 	}
 
 	ngOnInit() {
-		let self = this;
+		const self = this;
 		self._route.params.subscribe((params) => {
-			self.userId = params['userId'] * 1;
-			console.log("userId ==> " + self.userId);
+			self.userId = params.userId * 1;
+			console.log('userId ==> ' + self.userId);
 
 			self._userTechProfileModel.init(self.userId);
 
 			self._userService.getUserById(self.userId).then((data) => {
 				self.user = data;
-			})
+			});
 
-			self._functionPromiseService.initFunc(self.funcKey, () => {
-				return new Promise((resolve, reject) => {
+			self._functionPromiseService.initFunc(self.funcKey, () => new Promise((resolve, reject) => {
 					self._userTechProfileModel.waitingPromise().then(() => {
 						resolve({
-							getEnv: () => {
-								return environment;
-							},
-							getColorMeaningString: () => {
-								return "lightblue means this user's score says they have demonstrated skills at this level, and below. White, user has not been tested, or has not passed, at this level."
-							},
+							getEnv: () => environment,
+							getColorMeaningString: () => 'lightblue means this user\'s score says they have demonstrated skills at this level, and below. White, user has not been tested, or has not passed, at this level.',
 							getBackgroundColor: (lineItem, idx) => {
-								let score = self._userTechProfileModel.getScore(lineItem['id']);
-								
-								if (score == undefined) return "white";
-								if (score >= idx) return "lightblue"; else return "white";
+								const score = self._userTechProfileModel.getScore(lineItem.id);
+
+								if (score == undefined) {return 'white';}
+								if (score >= idx) {return 'lightblue';} else {return 'white';}
 							},
 							onLxDescriptionClick: (lineItem, idx) => {
-								self._router.navigate(['/line-item-action-page/' + self.userId + '/' + lineItem['id'] + '/' + idx]);
+								self._router.navigate(['/line-item-action-page/' + self.userId + '/' + lineItem.id + '/' + idx]);
 							}
-						})
-					})
-				})
-			})
-		})
+						});
+					});
+				}));
+		});
 	}
 
 	getDtimTechprofileComponentController() {
@@ -79,11 +73,11 @@ export class UserTechProfilePage implements OnInit {
 	}
 
 	getUserName() {
-		return this.user && this.user["name"];
+		return this.user && this.user.name;
 	}
 
 	getScore(lineItemId) {
-		let self = this;
+		const self = this;
 		return self._userTechProfileModel.getScore(lineItemId);
 	}
 
